a slightly evil commit
thru the magic of just pushing the problem into JavaScript and having it come back out typed we resolve the afore-mentioned problem of saved transitions having their types fixed.
Basically,  casting. But then it is an accurate model of what D3 is doing and the desire is not to make a strongly typed D3 but to enable D3 modules and idioms to be available in Purescript programs