drag zoom example works on Chrome
but doesn’t work on Safari? or maybe Safari is lying about not caching? but that could hardly be it, i’m stepping thru the code in a debugger, i’d notice if it wasn’t update. at any rate, f*** safari, chrome is now the reference browser for this project even if it’s energy use is pessimal and i have grave doubts about googles attitudes to privacy…